Remove barrel export for utils because it causes webpack errors in dev mode
