+++ src/components/content-staking/mutations/lockOrIncreaseTx.tsx
@@ -6,7 +6,10 @@ import {
   getGeneralEraInfoQuery,
 } from '@/services/contentStaking/generalErainfo/query'
 import { getCurrentWallet } from '@/services/subsocial/hooks'
-import { Status, createMutationWrapper } from '@/services/subsocial/utils'
+import {
+  Status,
+  createMutationWrapper,
+} from '@/services/subsocial/utils/mutation'
 import { getBalancesQuery } from '@/services/substrateBalances/query'
 import { useSendEvent } from '@/stores/analytics'
 import { useMyAccount } from '@/stores/my-account'
