+++ src/components/content-staking/mutations/unlockTx.tsx
@@ -11,7 +11,10 @@ import {
 } from '@/services/contentStaking/generalErainfo/query'
 import { getStakingConstsData } from '@/services/contentStaking/stakingConsts/query'
 import { getCurrentWallet } from '@/services/subsocial/hooks'
-import { Status, createMutationWrapper } from '@/services/subsocial/utils'
+import {
+  Status,
+  createMutationWrapper,
+} from '@/services/subsocial/utils/mutation'
 import { getBalancesQuery } from '@/services/substrateBalances/query'
 import { useSendEvent } from '@/stores/analytics'
 import { useMyAccount, useMyMainAddress } from '@/stores/my-account'
