+++ src/components/extensions/donate/DonateModal/donateForm/mutation.tsx
@@ -2,7 +2,7 @@ import Toast from '@/components/Toast'
 import useConnectWallet from '@/hooks/useConnectWallet'
 import { getChainsInfoQuery } from '@/services/chainsInfo/query'
 import { getCurrentWallet } from '@/services/subsocial/hooks'
-import { createMutationWrapper } from '@/services/subsocial/utils'
+import { createMutationWrapper } from '@/services/subsocial/utils/mutation'
 import { getBalancesQuery } from '@/services/substrateBalances/query'
 import { useMyAccount, useMyMainAddress } from '@/stores/my-account'
 import { useLazySubstrateMutation } from '@/subsocial-query/subsocial/lazyMutation'
