--- src/services/subsocial/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './mutation'
-export * from './service-mapper'
